requirements.txt
# Django and Django REST Framework
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0